refactor ride price functions into a static class
